refactor: Pluralizer
