+++ src/Complexes/Show/Carousel.jsx
@@ -3,6 +3,7 @@ import { Grid } from 'react-flexbox-grid';
 import styled from 'styled-components';
 
 import { getImageUrl } from '../../utils';
+import Pluralizer from '../../Pluralizer';
 
 const Carousel = styled.div`
   height: 400px;
@@ -35,21 +36,19 @@ const Button = styled.button`
   color: #fff;
 `;
 
-function declOfNum(number, titles) {
-  const cases = [2, 0, 1, 1, 1, 2];
-  const decl =
-    titles[number % 100 > 4 && number % 100 < 20 ? 2 : cases[number % 10 < 5 ? number % 10 : 5]];
-  return `${number} ${decl}`;
-}
-
 export default props =>
   (<Carousel>
     <Wrapper>
-      {props.imageList.map(data => <Photo src={getImageUrl(data.id, 512)} key={data.id} />)}
+      {props.images.map(data => <Photo src={getImageUrl(data.id, 512)} key={data.id} />)}
     </Wrapper>
     <Grid>
       <Button>
-        {declOfNum(props.imageList.length, ['фотография', 'фотографии', 'фотографий'])}
+        <Pluralizer
+          number={props.images.length}
+          one="фотография"
+          few="фотографии"
+          other="фотографий"
+        />
       </Button>
     </Grid>
   </Carousel>);
